nginx-ingress-controller - custom error backend
This commit introduces the software necessary to serve custom error
pages through the nginx ingress controller. Specifically, it includes
the actual server, the templates and the static resources associated
with the errors as well as the necessary configuration in term of
Kubernetes resources.

Additionally, it also drops the custom error pages from the frontend,
since they are now automatically handled by the ingress controller.