PR owner fetch
Before this commit, the owner of the PR was considered as the entity that applied the last modification.
Now the owner of the branch is the author of the first commit of the branch related to the PR. This is needed by the
dev event dispatcher, that triggers a new deployment creation through CrownOps in a namespace related to the PR owner.