Webservice: remove the k8s_library directory
This commit represents a preparation step to migrate to the kubernetes
javascript client library hosted on an external repository. Warning:
the state represented by this commit is NOT valid and the webservice
does not build correctly. Yet, this commit is kept separated from the
subsequend one for clarity, given the huge number of files removed.