Operators: prevent bad Ginkgo failures during panic
In case a panic happens during a test inside a goroutine, Ginkgo fails
to recover it and shows the complete stack trace instead of providing
useful information. This problem happens if a panic occurs inside the
Reconcile function, since it is executed using a goroutine.
Specifically, this situation may be triggered when gomock is used.
To solve this, it is necessary to call 'defer GinkgoRecover() inside
the goroutine. This commit provides a solution, introducing the
possibility to specify a function to be deferred at the beginning of the
Reconcile function, so that is can be configured only in case of tests.

Co-authored-by: alacuku <aldo.lacu@live.it>