instance-controller: remove oauth2-proxy creation
This commit removes the creation of the oauth2-proxy objects from the
instance operator. It assumes the existence of a centralized
oauth2-proxy deployment and configures the appropriate ingress
annotations based on the URL specified as input parameter.
Finally, it removes a set of CORS-related annotation which appear to
be no longer required.